+++ src/collective/exportimport/tests/test_setup.py
@@ -1,0 +1,33 @@
+# -*- coding: utf-8 -*-
+"""Setup tests for this package."""
+from plone import api
+from plone.app.testing import setRoles
+from plone.app.testing import TEST_USER_ID
+from collective.exportimport.testing import COLLECTIVE_EXPORTIMPORT_INTEGRATION_TESTING  # noqa: E501
+
+import unittest
+
+
+try:
+    from Products.CMFPlone.utils import get_installer
+except ImportError:
+    get_installer = None
+
+
+class TestSetup(unittest.TestCase):
+    """Test that collective.exportimport is properly installed."""
+
+    layer = COLLECTIVE_EXPORTIMPORT_INTEGRATION_TESTING
+
+    def setUp(self):
+        """Custom shared utility setup for tests."""
+        self.portal = self.layer['portal']
+        if get_installer:
+            self.installer = get_installer(self.portal, self.layer['request'])
+        else:
+            self.installer = api.portal.get_tool('portal_quickinstaller')
+
+    def test_restapi_installed(self):
+        """Test if restapi is installed, because we need it."""
+        self.assertTrue(self.installer.isProductInstalled(
+            'plone.restapi'))